Sweeping changes to support time input.
Adds a time panel when timeEnabled is set to true. The picker now only closes after choosing a time. When navigating with keyboard, enter/space moves focus over to the time picker, which can then be navigating with the up/down arrow keys.

Still room for improvement probably but I think it's at the stage where it could be a replacement for existing uses of non-accessible date/time pickers.